refactor: recent transaction screen migration to compose
