+++ app/src/main/java/org/mifos/mobile/ui/activities/HomeActivity.kt
@@ -39,7 +39,7 @@ import org.mifos.mobile.ui.getThemeAttributeColor
 import org.mifos.mobile.ui.help.HelpActivity
 import org.mifos.mobile.ui.home.HomeOldFragment
 import org.mifos.mobile.ui.login.LoginActivity
-import org.mifos.mobile.ui.recent_transactions.RecentTransactionsFragment
+import org.mifos.mobile.ui.recent_transactions.RecentTransactionsComposeFragment
 import org.mifos.mobile.utils.Constants
 import org.mifos.mobile.utils.TextDrawable
 import org.mifos.mobile.utils.Toaster
@@ -182,7 +182,7 @@ class HomeActivity :
             }
 
             R.id.item_recent_transactions -> replaceFragment(
-                RecentTransactionsFragment.newInstance(),
+                RecentTransactionsComposeFragment.newInstance(),
                 true,
                 R.id.container,
             )
@@ -407,7 +407,7 @@ class HomeActivity :
                     setNavigationViewSelectedItem(R.id.item_accounts)
                 }
 
-                is RecentTransactionsFragment -> {
+                is RecentTransactionsComposeFragment -> {
                     setNavigationViewSelectedItem(R.id.item_recent_transactions)
                 }
